+++ CHANGES.md
@@ -18,10 +18,10 @@ Features:
 -
 
 Changes:
--
+- OpinionSite can now return a "content" key holding downloaded content #873
 
 Fixes:
--
+- Fix `bva` scraper #873
 
 ## 2.8.1 - 2026-02-26